allow compilation of bzimage pkg under GOOS=tamago
